+++ minecraft/src/net/minecraft/client/gui/screen/ingame/CartographyTableScreen.java
@@ -87,7 +87,7 @@ public class CartographyTableScreen extends HandledScreen<CartographyTableScreen
 			this.drawMap(matrices, mapId, mapState, i + 86, j + 16, 0.34F);
 			RenderSystem.setShaderTexture(0, TEXTURE);
 			matrices.push();
-			matrices.translate(0.0, 0.0, 1.0);
+			matrices.translate(0.0F, 0.0F, 1.0F);
 			this.drawTexture(matrices, i + 67, j + 13 + 16, this.backgroundWidth, 132, 50, 66);
 			this.drawMap(matrices, mapId, mapState, i + 70, j + 32, 0.34F);
 			matrices.pop();
@@ -96,7 +96,7 @@ public class CartographyTableScreen extends HandledScreen<CartographyTableScreen
 			this.drawMap(matrices, mapId, mapState, i + 71, j + 17, 0.45F);
 			RenderSystem.setShaderTexture(0, TEXTURE);
 			matrices.push();
-			matrices.translate(0.0, 0.0, 1.0);
+			matrices.translate(0.0F, 0.0F, 1.0F);
 			this.drawTexture(matrices, i + 66, j + 12, 0, this.backgroundHeight, 66, 66);
 			matrices.pop();
 		} else {
@@ -108,7 +108,7 @@ public class CartographyTableScreen extends HandledScreen<CartographyTableScreen
 	private void drawMap(MatrixStack matrices, @Nullable Integer mapId, @Nullable MapState mapState, int x, int y, float scale) {
 		if (mapId != null && mapState != null) {
 			matrices.push();
-			matrices.translate((double)x, (double)y, 1.0);
+			matrices.translate((float)x, (float)y, 1.0F);
 			matrices.scale(scale, scale, 1.0F);
 			VertexConsumerProvider.Immediate immediate = VertexConsumerProvider.immediate(Tessellator.getInstance().getBuffer());
 			this.client.gameRenderer.getMapRenderer().draw(matrices, immediate, mapId, mapState, true, LightmapTextureManager.MAX_LIGHT_COORDINATE);
